refactor: simplify factory methods for DatabaseConfig and enforce validation (#688)
