+++ src/silo/test/float_equals_and_between.test.cpp
@@ -47,7 +47,7 @@ const std::vector<nlohmann::json> DATA = {
    createDataWithFloatNullValue("id_4")
 };
 
-const auto DATABASE_CONFIG = silo::config::DatabaseConfigReader().parseYaml(
+const auto DATABASE_CONFIG =
    R"(
 defaultNucleotideSequence: "segment1"
 schema:
@@ -58,9 +58,7 @@ schema:
     - name: "float_value"
       type: "float"
   primaryKey: "primaryKey"
-)"
-
-);
+)";
 
 const auto REFERENCE_GENOMES = ReferenceGenomes{
    {{"segment1", "A"}},
