+++ src/silo/test/insertion_contains.test.cpp
@@ -32,7 +32,7 @@ const std::vector<nlohmann::json> DATA = {
    createDataWithNucleotideInsertions("id_3", {{"segment1", {"123:CCC"}}, {"segment2", {}}}),
 };
 
-const auto DATABASE_CONFIG = silo::config::DatabaseConfigReader().parseYaml(
+const auto DATABASE_CONFIG =
    R"(
 defaultNucleotideSequence: "segment1"
 schema:
@@ -41,8 +41,7 @@ schema:
     - name: "primaryKey"
       type: "string"
   primaryKey: "primaryKey"
-)"
-);
+)";
 
 const auto REFERENCE_GENOMES = ReferenceGenomes{
    {{"segment1", "A"}, {"segment2", "T"}},
